JSON translation, now feat. QJsonValueRef
Back when the old JSON translation code was written, JSON documents were read-only, so we had to do a rather inefficient recursive rewrite of the whole document, creating a near-identical copy of it.

Now, we have QJsonValueRef, which lets us iterate over a JSON object/array and rewrite it in-place.